refactor to toga_helpers
